requirementsML.txt
tensorflow>=2.6.0
numpy>=1.19.5
Pillow>=8.0.0